+++ stand-alone/expressAdapter.js
@@ -38,7 +38,6 @@ function restOpFromRequest(req) {
         restOp.setHeader(headerName, req.headers[headerName]);
     });
 
-
     return restOp;
 }
 
@@ -55,7 +54,6 @@ function setResponseFromRestOp(restOp, res) {
         .status(restOp.getStatusCode())
         .send(body);
 }
-
 
 function getWorkerResponse(worker, req, res) {
     const restOp = restOpFromRequest(req);
@@ -82,6 +80,8 @@ function getWorkerResponse(worker, req, res) {
 }
 
 function generateApp(workers, options) {
+    options = options || {};
+
     if (!Array.isArray(workers)) {
         workers = [workers];
     }
@@ -115,10 +115,11 @@ function generateApp(workers, options) {
 
     // Create an endpoint to forward remaining requests to BIG-IP
     if (options.bigip) {
+        options.bigip.username = options.bigip.username || options.bigip.user;
         const endpoint = axios.create({
             baseURL: options.bigip.host,
             auth: {
-                username: options.bigip.user,
+                username: options.bigip.username,
                 password: options.bigip.password
             },
             maxBodyLength: 'Infinity',