+++ CHANGELOG.md
@@ -1,12 +1,23 @@
+# 1.14.0
+## Added
+
+## Fixed
+* Fix error and avoid delays when deleting or adding new templatesets
+* Fix error when updating bigip-fast-templates apps that use old pool_member definitions
+## Changed
+
 # 1.13.0
 ## Added
 * bigip-fast-templates: option for chain file on TLS Server in Exchange template
+* ipam: Add support for authorization headers other than basic auth
 
 ## Fixed
 * bigip-fast-templates: Fix SameSite=None cookies with a new iRule that adds the secure attribute to all cookies in the Exchange template 
 * bigip-fast-templates: Fix fqdn missing from Exchange services when not using the Common VIP
+* bigip-fast-templates: Fix the default port for pool members in the LDAP template, and show the password field
 
 ## Changed
+* Update f5-fast-core from v0.15.0 to v0.16.0
 
 # 1.12.0
 ## Added